Extract app↔daemon XPC contract into a WikiDaemonContract module
The XPC contract (the two @objc protocols + the transport error) was smeared
across WikiFSCore (protocols, buried in a 131-file module) and WikiCtlCore
(WikiDaemonError, mixed into the client transport). There was no single place
naming the app↔daemon boundary.

New Foundation-only leaf module WikiDaemonContract holds exactly the contract:
- WikiDaemonProtocol + WikiDaemonEventSink (git mv'd from WikiFSCore/Core/)
- WikiDaemonError (extracted from WikiCtlCore/WikiDaemonConnection.swift)
- Contract.swift — module doc explaining the boundary

No DTOs move: every payload crosses the wire JSON-encoded as Data, so the @objc
protocol signatures reference no domain types. The contract is a true leaf with
zero domain coupling; payload DTOs (WikiDescriptor, QueueItemRequest, AgentEvent,
chat/queue DTOs) stay in WikiFSCore/WikiFSEngine, encoded/decoded by the typed
client (WikiCtlCore) and server (wikid) wrappers.

Edges: WikiDaemonContract ← { wikid (server), WikiFS (app), WikiCtlCore (client),
WikiFSAppTests }. WikiFSCore/WikiFSEngine do NOT depend on it (their references
are doc comments). All source is #if os(macOS) — empty on Linux, edges harmless.

swift build ✓ (all products link); 56 daemon/contract tests ✓; WikiFSCore clean
of the moved symbols.
